requirements.txt
flask
sqlalchemy
flask_restful
flask_cors
PyJWT